[MCJ-1410] refactor: 이메일 마스킹 로직 공통 유틸로 분리
